requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
SQLAlchemy>=2.0.0
Alembic>=1.12.0
psycopg[binary]>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
pyotp>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
email-validator>=2.0.0
Flask-Limiter>=3.5.0
itsdangerous>=2.1.0
Babel>=2.12.0
gunicorn>=21.0.0 